Fix no owners for subfolders
