+++ pytest.ini
@@ -1,3 +1,4 @@
 [pytest]
 addopts =
-    --cov-config=pyproject.toml --cov-report=html --cov-report=term --cov=github_webhook_server
+    --pdbcls=IPython.terminal.debugger:TerminalPdb
+    --cov-config=pyproject.toml --cov-report=html --cov-report=term --cov=webhook_server_container
